+++ src/evm/opcodes/delegatecall.ts
@@ -4,10 +4,10 @@ import { Address } from '../../address';
 import { EvmContext } from '../vm/evm-context';
 import { messageCall } from '../vm/message-call';
 
-class CallOp implements OpCode {
-  public readonly code = 0xf1;
-  public readonly mnemonic = 'CALL';
-  public readonly description = 'Message-call into an account';
+class DelegateCallOp implements OpCode {
+  public readonly code = 0xf4;
+  public readonly mnemonic = 'DELEGATECALL';
+  public readonly description = 'Message-call into this account but with alternatives account code';
   public readonly gas = 0;
   public readonly bytes = 1;
 
@@ -16,11 +16,21 @@ class CallOp implements OpCode {
   }
 
   public async handle(context: EvmContext) {
-    const { stack, worldState, memory, origin, executor, gasPrice, callDepth, modify } = context;
+    const {
+      stack,
+      worldState,
+      memory,
+      caller,
+      origin,
+      executor,
+      gasPrice,
+      callDepth,
+      modify,
+      executionValue,
+    } = context;
 
     const gas = stack.pop();
     const addr = stack.pop();
-    const value = stack.pop();
     const inOffset = stack.pop();
     const inSize = stack.pop();
     const retOffset = stack.pop();
@@ -31,14 +41,14 @@ class CallOp implements OpCode {
 
     const { txSubstrate, status, returned } = await messageCall(
       worldState,
-      executor,
+      caller,
       origin,
-      recipient,
+      executor,
       recipient,
       gas,
       gasPrice,
-      value,
-      value,
+      BigInt(0),
+      executionValue,
       calldata,
       callDepth + 1,
       modify,
@@ -59,4 +69,4 @@ class CallOp implements OpCode {
   }
 }
 
-export const Call = new CallOp();
+export const DelegateCall = new DelegateCallOp();