+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.5] - 2019-03-10
+
+### Added
+
+- `EvmProvider` opcodes and precompiles for contracts 6, 7 and 8, plus more.
+- `nonce` can be added to contract tx sends.
+- `decodeFunctionData` method added to `ContractAbi` to simplify manual tx decoding.
+
 ## [3.0.4] - 2019-03-02
 
 ### Changed
@@ -173,6 +181,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of Typescript port from web3.js.
 
+[3.0.5]: https://github.com/xf00f/web3x/compare/v3.0.4...v3.0.5
 [3.0.4]: https://github.com/xf00f/web3x/compare/v3.0.3...v3.0.4
 [3.0.3]: https://github.com/xf00f/web3x/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/xf00f/web3x/compare/v3.0.1...v3.0.2